+++ modules/template/elements/MarkdownElement.php
@@ -63,6 +63,14 @@ class MarkdownElement extends BaseElementContent implements \Stringable
     /**
      * @inheritdoc
      */
+    /**
+     * @inheritdoc
+     */
+    public static function isInlineEditingEnabledByDefault(): bool
+    {
+        return true;
+    }
+
     public function __toString(): string
     {
         return (string) Richtext::output($this->content);