+++ widgets/TinyMce.php
@@ -41,6 +41,7 @@ class TinyMce extends \dosamigos\tinymce\TinyMce
 
         $this->clientOptions = ArrayHelper::merge([
             'license_key' => 'gpl',
+            'promotion' => false,
             'plugins' => ['code', 'autolink', 'link', 'image', 'lists', 'fullscreen', 'table', 'wordcount', 'anchor', 'lists'],
             'menu' => ['insert' => [
                 'title' => Yii::t('CustomPagesModule.base', 'Insert'),